+++ test/page-ex.test.js
@@ -443,6 +443,30 @@ describe('Page extensions', function() {
       page.compile('Hello', 'md', 'index', { name: 'Tobi', pedantic: true });
     }); // should convert markup, render layout and write page with options
     
+    it('should convert markup, render layout and invoke callback', function(done) {
+      var app = new function(){};
+      app.convert = sinon.stub().returns('<p>Hello</p>');
+      app.render = sinon.stub().yieldsAsync(null, '<p>Hello</p>');
+      
+      var page = new Page();
+      setPrototypeOf(page, Object.create(pagex, {
+          app: { configurable: true, enumerable: true, writable: true, value: app }
+        }));
+      
+      page.compile('Hello', 'md', 'index', function(err, str) {
+        if (err) { return done(err); }
+        expect(str).to.equal('<p>Hello</p>');
+        expect(page.app.convert.getCall(0).args[0]).to.equal('Hello');
+        expect(page.app.convert.getCall(0).args[1]).to.equal('md');
+        expect(page.app.render.getCall(0).args[0]).to.equal('index');
+        expect(page.app.render.getCall(0).args[1]).to.deep.equal({
+          content: '<p>Hello</p>',
+          _locals: undefined
+        });
+        done();
+      });
+    }); // should convert markup, render layout and invoke callback
+    
   }); // #compile
   
 });