+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Express conventions.
 - Renamed `app#convert()` to `app#markup()`.
 - `app#convert` is now a synchronous function, returning a string rather than
 yielding an error and string to a callback function.
+- `page#compile` now invokes callback after rendering.  If content is needed
+without being inserted into a layout, call `page#convert`.
 
 ### Removed